requirements.txt
python-telegram-bot[job-queue]==21.7
httpx==0.27.2
python-dotenv==1.0.1
groq==0.13.0
pydantic==2.9.2
pydantic-settings==2.6.1
aiosqlite==0.20.0
eth-utils==5.1.0
eth-abi==5.1.0
apscheduler==3.10.4
pillow==11.0.0
